chore: slight dialog refactor
